agentic_AI/port_agent/requirements.txt
anthropic
python-dotenv
rich
jinja2
pyyaml
